Rename pipewire player
